webhook-mongo/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor==3.3.1
pydantic==2.3.0
python-dotenv==1.0.0
pyyaml==6.0.1
